feat(taskManager): add ModalContent component with form handling and inline editing
- Extract modal content logic into a reusable ModalContent component for better maintainability.
- Integrate react-hook-form with FormProvider for controlled form inputs and validation.
- Add support for AI assignment ("Cashport AI") and attachment reprocessing.
- Replace the previous implementation with a form-based approach using useWatch and Controller.
